+++ demo/src/index.js
@@ -169,6 +169,80 @@ class Example3 extends Component {
     }
 }
 
+class Example4 extends Component {
+
+    state = {
+        value: null
+    }
+
+    optionRenderer = ({option}) => (
+        <div
+            className={css`
+                    background-color: rgb(222, 54, 24);
+                    color: rgb(255, 255, 255);
+                    min-width: 18px;
+                    height: 18px;
+                    font-size: 13px;
+                    font-weight: 400;
+                    max-width: 100%;
+                    align-items: center;
+                    display: inline-flex;
+                    padding-left: 8px;
+                    padding-right: 8px;
+                    line-height: 1.5;
+                    -webkit-print-color-adjust: exact;
+                    border-radius: 9999px;
+                    flex: 0 0 auto;
+                `}
+        >
+            {option.name}
+        </div>
+    )
+
+    render() {
+        return <div>
+            <h2>
+                Custom optionRenderer
+            </h2>
+            <div
+                className={css`
+                    width: 200px;
+                    height: 40px;
+                    display: flex;
+                `}
+            >
+                <Select
+                    clearable={true}
+                    value={this.state.value}
+                    alignLeft={true}
+                    optionRenderer={this.optionRenderer}
+                    options={[{
+                        id: 'opt1',
+                        name: 'Option 1'
+                    }, {
+                        id: 'opt2',
+                        name: 'Option 2'
+                    }, {
+                        id: 'opt3',
+                        name: 'Option 3'
+                    }]}
+                    onChange={({value}) => {
+
+                        this.setState({
+                            value
+                        })
+                    }}
+                />
+            </div>
+            <h3>
+                State
+            </h3>
+            <pre>
+                {JSON.stringify(this.state, null, 2)}
+            </pre>
+        </div>
+    }
+}
 
 class Demo extends React.Component {
 
@@ -176,10 +250,11 @@ class Demo extends React.Component {
 
         return (
             <div>
-                <h1>SingleSelect Demo</h1>
+                <h1>Select Demo</h1>
                 <Example1/>
                 <Example2/>
                 <Example3/>
+                <Example4/>
             </div>
         )
     }