+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2019-01-27
+
+### Removed
+- Removed `optionNameGetter`. This can now be controlled using the `optionRenderer` prop.
+
+### Added
+- Added `optionRenderer` prop responsible for rendering the option.
+
 ## [0.2.0] - 2019-01-24
 ### Added
 - Support for clearing the choice.